requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
cfscrape==2.1.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
cssselect==1.2.0
Django==4.2.2
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
exceptiongroup==1.1.1
filelock==3.12.2
flower==1.2.0
generics==6.0.0
h11==0.14.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.3.1
lxml==4.9.2
outcome==1.2.0
packaging==23.1
parsel==1.8.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
queuelib==1.6.2
redis==4.5.5
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
Scrapy==2.9.0
selenium==4.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tldextract==3.4.4
tornado==6.3.2
trio==0.22.0
trio-websocket==0.10.3
Twisted==22.10.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
w3lib==2.1.1
wcwidth==0.2.6
wsproto==1.2.0
zope.interface==6.0
